% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loan}
\alias{loan}
\title{SCU loan and loan transaction data (from 2021 to 2022)}
\format{
\subsection{\code{loan}: A data frame with 185777 rows and 25 columns.}{

#' \describe{
\item{PARENTACCOUNT}{Member account number.}
\item{ID}{Identifying the Loan record.}
\item{LOANCODE}{The loan code defines certain characteristics of the loan and determines how Episys processes the loan.}
\item{CREDITSCORE}{Member account number.}
\item{ORIGINALBALANCE}{This field stores the original balance of the loan or the balance after an add-on to the loan.}
\item{ORIGINALRATE}{The original effective APR (Annual Percentage Rate) when the loan was opened.}
\item{TYPE}{Type of loan.}
\item{INTERESTRATE}{The interest rate charged for the loan.}
\item{BRANCH}{The branch that ""owns"" the loan.}
\item{PURPOSECODE}{Purpose of the loan.}
\item{APPROVALDATE}{The date the loan was approved.}
\item{PAYMENT}{The amount of the standard loan payment.}
\item{DQ}{Target variable.}
\item{BALANCE_01}{The current principal balance of the loan in month_1}
\item{BALANCE_02}{The current principal balance of the loan in month_2}
\item{BALANCE_03}{The current principal balance of the loan in month_3}
\item{BALANCE_04}{The current principal balance of the loan in month_4}
\item{BALANCE_05}{The current principal balance of the loan in month_5}
\item{BALANCE_06}{The current principal balance of the loan in month_6}
\item{BALANCE_07}{The current principal balance of the loan in month_7}
\item{BALANCE_08}{The current principal balance of the loan in month_8}
\item{BALANCE_09}{The current principal balance of the loan in month_9}
\item{BALANCE_10}{The current principal balance of the loan in month_10}
\item{BALANCE_11}{The current principal balance of the loan in month_11}
\item{BALANCE_12}{The current principal balance of the loan in month_12}
}
}
}
\usage{
loan
}
\description{
Contains information of loan and transactions.
}
\details{
\enumerate{
\item loan data
}
}
\keyword{datasets}
